refactor(admin): extract CertificationsTab into its own file
